[UPD] Load Token Address in main page
